[zlib] Move windows specific code to XXXXX_win.cpp files
