+++ VS2017/ZLibrary.vcxproj
@@ -114,10 +114,10 @@
     </Lib>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="..\src\zlib\base.cpp" />
+    <ClCompile Include="..\src\zlib\base_win.cpp" />
     <ClCompile Include="..\src\zlib\color.cpp" />
     <ClCompile Include="..\src\zlib\event.cpp" />
-    <ClCompile Include="..\src\zlib\FTPSession.cpp" />
+    <ClCompile Include="..\src\zlib\FTPSession_win.cpp" />
     <ClCompile Include="..\src\zlib\input.cpp" />
     <ClCompile Include="..\src\zlib\matrix.cpp" />
     <ClCompile Include="..\src\zlib\matrix2.cpp" />
